### 2021-01-04 #2937 @arfon
:wave: @kthyng @leouieda - would one of you be willing to handle this submission for JOSS? @kbarnhart is the suggested editor but is already handing four submissions right now 😬.
