### 2021-01-04 #2937 @kbarnhart
@jonschwenk thanks for this submission. At this point I've looked through the repository and have determined that I can move forward to looking for reviewers.  If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, this [list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

In addition there are a few minor items I would recommend attending to before the formal review issue begins:
- [ ] add the identified missing DOI
- [ ] address the date format of Tejedor et al., 2019 (line 46)
- [ ] add references for dependency software where citations exists.
