### 2021-01-05 6308a8d
fix figures in paper
