+++ paper.bib
@@ -10131,6 +10131,25 @@
   number = {2}
 }
 
+@article{schwenk2017b,
+  title = {High Spatiotemporal Resolution of River Planform Dynamics from {{Landsat}}: {{The RivMAP}} Toolbox and Results from the {{Ucayali River}}},
+  shorttitle = {High Spatiotemporal Resolution of River Planform Dynamics from {{Landsat}}},
+  author = {Schwenk, Jon and Khandelwal, Ankush and Fratkin, Mulu and Kumar, Vipin and Foufoula-Georgiou, Efi},
+  year = {2017},
+  volume = {4},
+  pages = {46--75},
+  issn = {2333-5084},
+  doi = {10.1002/2016EA000196},
+  abstract = {Quantifying planform changes of large and actively migrating rivers such as those in the tropical Amazon at multidecadal time scales, over large spatial domains, and with high spatiotemporal frequency is essential for advancing river morphodynamic theory, identifying controls on migration, and understanding the roles of climate and human influences on planform adjustments. This paper addresses the challenges of quantifying river planform changes from annual channel masks derived from Landsat imagery and introduces a set of efficient methods to map and measure changes in channel widths, the locations and rates of migration, accretion and erosion, and the space-time characteristics of cutoff dynamics. The techniques are assembled in a comprehensive MATLAB toolbox called RivMAP (River Morphodynamics from Analysis of Planforms), which is applied to over 1500 km of the actively migrating and predominately meandering Ucayali River in Peru from 1985 to 2015. We find multiscale spatial and temporal variability around multidecadal trends in migration rates, erosion and accretion, and channel widths revealing a river dynamically adjusting to sediment and water fluxes. Confounding factors controlling planform morphodynamics including local inputs of sediment, cutoffs, and climate are parsed through the high temporal analysis.},
+  annotation = {\_eprint: https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/2016EA000196},
+  copyright = {\textcopyright 2016. The Authors.},
+  file = {C\:\\Users\\Jon\\Zotero\\storage\\2HUI494N\\Schwenk et al. - 2017 - High spatiotemporal resolution of river planform d.pdf;C\:\\Users\\Jon\\Zotero\\storage\\8PIQU3U4\\2016EA000196.html},
+  journal = {Earth and Space Science},
+  keywords = {image processing,Landsat,MATLAB,meander dynamics,planform change,river migration},
+  language = {en},
+  number = {2}
+}
+
 @inproceedings{schwenk2018,
   title = {Automatic {{Extraction}} of {{Channel Network Topology}} ({{RivGraph}})},
   booktitle = {{{AGU Fall Meeting Abstracts}}},
