### 2021-01-05 0e06e43
figure formatting
