+++ paper.md
@@ -29,7 +29,7 @@ River networks sustain life and landscapes by carrying and distributing water, s
 
 `RivGraph` is a Python package that automates the extraction and characterization of river channel networks from a user-provided binary image, or mask, of a channel network (Fig. \autoref{fig:corefunctions}). Masks may be derived from (typically remotely-sensed) imagery, simulations, or even hand-drawn. `RivGraph` will create explicit representations of the channel network by resolving river centerlines as links, and junctions as nodes. Flow directions are solved for each link of the network without using auxiliary data, e.g. a digital elevation model (DEM). Morphologic properties are computed as well, including link lengths, widths, sinuosities, branching angles, and braiding indices. If provided, `RivGraph` will preserve georeferencing information of the mask and will export results as ESRI shapefiles, GeoJSONs, and GeoTIFFs for easy import into GIS software. `RivGraph` can also return extracted networks as `networkx` objects for convenient interfacing with the full-featured `networkx` package [@hagberg2008]. Finally, `RivGraph` offers a suite of topologic metrics that were specifically designed for river channel network analysis [@tejedor2015].
 
-![The core functionality of RivGraph for a delta channel network.\label{fig:corefunctions}](https://github.com/jonschwenk/RivGraph/blob/master/examples/images/rivgraph_overview_white.PNG)
+![The core functionality of RivGraph for a delta channel network.\label{fig:corefunctions}](https://github.com/jonschwenk/RivGraph/blob/master/examples/images/rivgraph_overview_white.jpg)
 
 # Statement of need
 
@@ -58,7 +58,7 @@ Braided rivers should be analyzed with the `river` class, which instead of a use
 - `compute_centerline()` : computes the centerline of the holes-filled river mask (not individual channels)
 - `compute_mesh()` : creates a mesh of evenly-spaced transects that are approximately perpendicular to the centerline. The user can specify the mesh spacing and transect width.
 
-![A RivGraph-generated mesh for a mask of the Indus River.\label{fig:indusmesh}](https://github.com/jonschwenk/RivGraph/blob/master/examples/images/indus_mesh_paper.PNG)
+![A RivGraph-generated mesh for a mask of the Indus River.\label{fig:indusmesh}](https://github.com/jonschwenk/RivGraph/blob/master/examples/images/indus_mesh_paper.JPG)
 
 ### Advanced Functionality
 
