### 2021-01-05 0029a2d
figure formatting
