### 2021-01-05 92cf5d6
relative image paths in paper
