### 2021-01-05 1f97128
figure extension type
