### 2021-01-05 355548b
figure width in paper
