### 2021-01-07 #2937 @kbarnhart
@jonschwenk thanks for pointing me to the test instructions. That is sufficient. I had managed to miss that. 🤦

Regarding citations. One of the review criteria in the checklist that reviewers will use states

> **References**: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

Thus I recommend adding citations to the paper to reference software used to construct RivGraph.

Let me know if that doesn't address your question.

Also, please let me know if you have recommendations for reviewers. I have some thoughts on suitable reviewers for this, but it is always helpful to get input from authors.
