### 2021-01-08 cfb9ea8
add dependency citations to paper
