+++ paper.bib
@@ -1536,6 +1536,17 @@
   number = {3}
 }
 
+@article{bradski2000,
+  title = {The {{OpenCV}} Library},
+  author = {Bradski, G.},
+  year = {2000},
+  citeulike-article-id = {2236121},
+  journal = {Dr. Dobb's Journal of Software Tools},
+  keywords = {bibtex-import},
+  posted-at = {2008-01-15 19:21:54},
+  priority = {4}
+}
+
 @article{brady2013,
   title = {Modelling Adult {{Aedes}} Aegypti and {{Aedes}} Albopictus Survival at Different Temperatures in Laboratory and Field Settings},
   author = {Brady, Oliver J. and Johansson, Michael A. and Guerra, Carlos A. and Bhatt, Samir and Golding, Nick and Pigott, David M. and Delatte, H{\'e}l{\`e}ne and Grech, Marta G. and Leisnham, Paul T. and {Maciel-de-Freitas}, Rafael},
@@ -4012,6 +4023,14 @@
   address = {{New Orleans}}
 }
 
+@manual{gdal2020,
+  title = {{{GDAL}}/{{OGR}} Geospatial Data Abstraction Software Library},
+  author = {{GDAL/OGR contributors}},
+  year = {2020},
+  organization = {{Open Source Geospatial Foundation}},
+  type = {Manual}
+}
+
 @article{geiger2013,
   title = {Satellite-Derived Coastal Ocean and Estuarine Salinity in the {{Mid}}-{{Atlantic}}},
   author = {Geiger, Erick F. and Grossi, Matthew D. and Trembanis, Arthur C. and Kohut, Josh T. and Oliver, Matthew J.},
@@ -4100,6 +4119,20 @@
   keywords = {bayesian computation,gibbs sampling,markov chain monte carlo method,maximum-likelihood-estimation,metropolis algorithm,models,pharmacokinetic model,population pharmacokinetic parameters,random variate generation},
   language = {English},
   number = {4}
+}
+
+@article{gillies2007,
+  title = {Shapely: Manipulation and Analysis of Geometric Objects},
+  author = {Gillies, Sean and others},
+  year = {2007},
+  organization = {{toblerity.org}}
+}
+
+@article{gillies2011,
+  title = {Fiona Is {{OGR}}'s Neat, Nimble, No-Nonsense {{API}}},
+  author = {Gillies, Sean and others},
+  year = {2011},
+  organization = {{Toblerity}}
 }
 
 @article{gilvear2000,
@@ -4855,6 +4888,24 @@
   number = {4}
 }
 
+@article{harris2020,
+  title = {Array Programming with {{NumPy}}},
+  author = {Harris, Charles R. and Millman, K. Jarrod and {van der Walt}, St{\'e}fan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and {van Kerkwijk}, Marten H. and Brett, Matthew and Haldane, Allan and {del R{\'i}o}, Jaime Fern{\'a}ndez and Wiebe, Mark and Peterson, Pearu and {G{\'e}rard-Marchant}, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
+  year = {2020},
+  month = sep,
+  volume = {585},
+  pages = {357--362},
+  publisher = {{Nature Publishing Group}},
+  issn = {1476-4687},
+  doi = {10.1038/s41586-020-2649-2},
[…]
