### 2021-01-10 #2937 @jonschwenk
Ok, I didn't see a way to organically work the dependency refs into the paper so I just added a new section for that.

Here are some possible reviewers:
Erik Nesvold - https://github.com/eriknes
Charlie Marshak (author of "competing" package Orinoco) - https://scienceandtechnology.jpl.nasa.gov/charlie-marshak
Allison Pfeiffer or Jon Czuba

From JOSS list (Python preferred, 'network' or 'geomorphology' in interests):
RafalKucharskiPK/
jkbren
leotrs
nicgaspar
marda-science and dbuscombe-usgs

Hope that helps!
