### 2021-01-13 #2937 @kbarnhart
@leotrs and @eriknes thanks for your willingness to review. I'll now add you as reviewers and start the review.

@cmarshak if you are interested in reviewing I can add you after I start the review.

@dbuscombe-usgs and @pfeiffea thanks for the prompt replies.
