### 2021-01-13 #2952 @whedon
[REVIEW]: RivGraph: Automatic extraction and analysis of river and delta channel network topology
**Submitting author:** @jonschwenk (<a href="http://orcid.org/0000-0001-5803-9686">Jon Schwenk</a>)
**Repository:** <a href="https://github.com/jonschwenk/RivGraph" target ="_blank">https://github.com/jonschwenk/RivGraph</a>
**Version:** v0.4
