### 2021-01-13 #2952 @jonschwenk
@kbarnhart After a Zotero update, it started exporting my full library by default instead of just a particular collection. whedon's DOI check is thus of my entire library, not just the references in the paper. I have fixed the .bib file in the RivGraph repo to only contain the relevant refs, but I'm not sure how to ask whedon to re-run the DOI check.
