### 2021-01-13 #2952 @kbarnhart
@jonschwenk done. for reference [here is the list of things whedon can do](https://joss.readthedocs.io/en/latest/whedon.html)
