### 2021-01-25 #2952 @leotrs
In trying to install the package, I ran into some problems reported [here](https://github.com/jonschwenk/RivGraph/issues/38). I will continue my review pending the fix.
