+++ requirements.txt
@@ -1,11 +1,0 @@
-numpy
-scikit-image
-opencv
-pandas
-networkx
-geopandas
-pyproj
-Shapely
-Fiona
-matplotlib
-fastdtw
