### 2021-01-25 jonschwenk/RivGraph#38 @jonschwenk
Thanks for the report, @leotrs.

The requirements.txt file was outdated--apologies. We have removed it entirely, and now only point now to the enviornment.yml file.

@elbeejay has updated the documentation (see PR #39 ) for installation from source to include links to each of the dependency repos--this should take care of the mismatched names (e.g. opencv).

As noted in that document, gdal is notoriously difficult sometimes to install properly. We can't really provide support for that, as the problems are too difficult to reproduce on different systems (and can be due to a number of things--paths seem to be one of the main culprits).

As for the conda installation, we use this method frequently and have not encountered the issue you report. The error message suggests something may be amiss in the .yml file, but again, we use this same file for our Github Actions builds without issue.

Perhaps one thing to try that might "kill two birds" would be to install miniconda (instead of Anaconda). It takes up far less space, and perhaps a fresh install might solve the conda install issue.

@elbeejay please chime in with any additional notes (or correct me where I'm wrong).
