### 2021-01-27 jonschwenk/RivGraph#38 @leotrs
Ok, I was finally able to install. The above error is totally my fault! Apologies.

Here's what happened in [the instructions](https://jonschwenk.github.io/RivGraph/install/index.html), it says to `(right-click, "Save As...")`. I understood that too literally and actually right-clicked _the link_ and chose "Save link as...". However, this saves the HTML source code of the page the link points to, so I actually saved the HTMl of [this page](https://github.com/jonschwenk/RivGraph/blob/master/environment.yml). I didn't realize this and just executed `$ conda env create --file /path/to/environment.yml`, so I was giving it the wrong file.

What I had to do was go to [this page](https://github.com/jonschwenk/RivGraph/blob/master/environment.yml), click on "Raw", and copy&save the contents manually to an `environment.yml` file. And now `$ conda env create --file /path/to/environment.yml` works perfectly. This is with a fresh installation of miniconda, though I'm sure it would also have worked with my previous attempt using conda.

BTW, this is on Chrome/Ubuntu.
