### 2021-01-27 a0e7fa0
adding connectivity option for islands
