### 2021-01-27 jonschwenk/RivGraph#38 @jonschwenk
Ah, ok that makes sense. Our instructions could be confusing--I will change the ("right-click, "Save As...") to a safer instruction. Glad you were able to sort it out!
