### 2021-01-27 2578a35
update install instructions
