### 2021-01-27 jonschwenk/RivGraph#38 @kbarnhart
linking with the JOSS review issue

openjournals/joss-reviews/issues/2952
