### 2021-01-27 #2952 @kbarnhart
@eriknes and @leotrs, we have an automatic reminder two weeks into the review. Feel free to provide an update if you have one.

@leotrs making an issue in the main repo and linking it here is perfect. Thank you! If you paste `openjournals/joss-reviews/issues/2952` into the issue then GitHub links the issues nicely.
