### 2021-01-28 jonschwenk/RivGraph#40 @leotrs
Link in the README leads to 404
At the end of the About section in the README, there is link to a "delta channel network" that links to https://github.com/jonschwenk/RivGraph/blob/master/examples/delta_example.py.ipynb. This leads to a 404 on my end.
