### 2021-01-28 jonschwenk/RivGraph#40 @leotrs
I just realized that the extension in the link is wrong (.py.ipynb instead of just .ipynb)
