### 2021-01-28 jonschwenk/RivGraph#41 @leotrs
Broken link in documentation
On the first page of the documentation, there is a broken link to the FAQ:

![2021-01-28-093631_793x55_scrot](https://user-images.githubusercontent.com/1096704/106153234-523bed80-614c-11eb-8aeb-51778853173c.png)
