### 2021-01-28 jonschwenk/RivGraph#40 @jonschwenk
Good eye, [fixed](https://github.com/jonschwenk/RivGraph/commit/10d44220690db51f67c18dd20db88c9f3b5c50e9).
