### 2021-01-28 jonschwenk/RivGraph#42 @leotrs
Can't run example notebook
I'm running the delta_example notebook. Everything runs well until Section 4. Pruning the network:

```python
colville.prune_network()
# Note that we can also specify the location of the shoreline and inlet nodes:
# colville.prune_network(path_shoreline='/path/to/shoreline/file', path_inletnodes='/path/to/inletnodes/file')

# Now that we've pruned, we should re-export the network:
colville.to_geovectors()
# Note that this time we didn't specify the arguments; by default 'network' will be exported as type 'json'.

CPLE_OpenFailedError                      Traceback (most recent call last)
fiona/_shim.pyx in fiona._shim.gdal_open_vector()

fiona/_err.pyx in fiona._err.exc_wrap_pointer()

CPLE_OpenFailedError: data/Colville_Delta/Results/Colville_inlet_nodes.shp: No such file or directory

DriverError                               Traceback (most recent call last)
<ipython-input-8-2181e62ecfce> in <module>
----> 1 colville.prune_network()
      2 # Note that we can also specify the location of the shoreline and inlet nodes:
      3 # colville.prune_network(path_shoreline='/path/to/shoreline/file', path_inletnodes='/path/to/inletnodes/file')
      4
      5 # Now that we've pruned, we should re-export the network:

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/rivgraph/classes.py in prune_network(self, path_shoreline, path_inletnodes)
    628             raise AttributeError('Could not inlet_nodes shapefile which should be at {}.'.format(self.paths['inlet_nodes']))
    629
--> 630         self.links, self.nodes = du.prune_delta(self.links, self.nodes, path_shoreline, path_inletnodes, self.gdobj)
    631
    632

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/rivgraph/deltas/delta_utils.py in prune_delta(links, nodes, shoreline_shp, inlets_shp, gdobj)
     45     """
     46     # Get inlet nodes
---> 47     nodes = find_inlet_nodes(nodes, inlets_shp, gdobj)
     48
     49     # Remove spurs from network (this includes valid inlets and outlets)

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/rivgraph/deltas/delta_utils.py in find_inlet_nodes(nodes, inlets_shp, gdobj)
    100
    101     # Check that CRSs match; reproject inlet points if not
--> 102     inlets_gpd = gpd.read_file(inlets_shp)
    103     mask_crs = CRS(gdobj.GetProjection())
    104     if inlets_gpd.crs != mask_crs:

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/geopandas/io/file.py in read_file(filename, bbox, mask, rows, **kwargs)
     87
     88     with fiona_env():
---> 89         with reader(path_or_bytes, **kwargs) as features:
     90
     91             # In a future Fiona release the crs attribute of features will

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/fiona/env.py in wrapper(*args, **kwargs)
    398     def wrapper(*args, **kwargs):
    399         if local._env:
--> 400             return f(*args, **kwargs)
    401         else:
    402             if isinstance(args[0], str):

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/fiona/__init__.py in open(fp, mode, driver, schema, crs, encoding, layer, vfs, enabled_drivers, crs_wkt, **kwargs)
    254
    255         if mode in ('a', 'r'):
--> 256             c = Collection(path, mode, driver=driver, encoding=encoding,
    257                            layer=layer, enabled_drivers=enabled_drivers, **kwargs)
    258         elif mode == 'w':

~/miniconda3/envs/rivgraph/lib/python3.8/site-packages/fiona/collection.py in __init__(self, path, mode, driver, schema, crs, encoding, layer, vsi, archive, enabled_drivers, crs_wkt, ignore_fields, ignore_geometry, **kwargs)
    160             if self.mode == 'r':
    161                 self.session = Session()
--> 162                 self.session.start(self, **kwargs)
    163             elif self.mode in ('a', 'w'):
    164                 self.session = WritingSession()

fiona/ogrext.pyx in fiona.ogrext.Session.start()

fiona/_shim.pyx in fiona._shim.gdal_open_vector()

DriverError: data/Colville_Delta/Results/Colville_inlet_nodes.shp: No such file or directory
```

I believe the correct file is `DriverError: data/Colville_Delta/Colville_inlet_nodes.shp`. I am running the notebook from within the `examples/` directory, and I haven't modified any of the code cells.

Did I do something wrong? Should I execute the notebook from a different directory? (If so, perhaps it would be useful to document which directory)
