### 2021-01-28 jonschwenk/RivGraph#42 @jonschwenk
No, this was not your error. We redesigned the default path management and failed to update the examples. In order to keep you going, you can specify the paths as keywords to the `prune_network()` argument:

`colville.prune_network(path_shoreline='data/Colville_Delta/Colville_shoreline.shp', path_inletnodes='data/Colville_Delta/Colville_inlet_nodes.shp)`

I will close this Issue once we've fixed the example.
