### 2021-01-28 jonschwenk/RivGraph#42 @jonschwenk
[Fixed](https://github.com/jonschwenk/RivGraph/commit/83e73ebdb6a7bb511c3dbff35966ad9ab5efcae0) the example. Re-ran all examples and found no errors.
