### 2021-02-01 #2952 @leotrs
I have finished the review checklist and it is my opinion that the software and manuscript meet JOSS standards.

Issues in target repository associated to my review:
+ https://github.com/jonschwenk/RivGraph/issues/38
+ https://github.com/jonschwenk/RivGraph/issues/40
+ https://github.com/jonschwenk/RivGraph/issues/41
+ https://github.com/jonschwenk/RivGraph/issues/42
