### 2021-02-07 jonschwenk/RivGraph#41 @kbarnhart
openjournals/joss-reviews/issues/2952
