### 2021-02-07 jonschwenk/RivGraph#42 @kbarnhart
openjournals/joss-reviews/issues/2952
