### 2021-02-17 caaaf86
flux-related fixes and doc update
