### 2021-02-25 #2952 @eriknes
Dear Katy,

It seems like the invitation may have expired - can you resend it to me? I
will complete the review tomorrow morning CET time.

I am sorry it has taken some time.

Best regards,

Erik

On Thu, Feb 18, 2021 at 2:16 PM Katy Barnhart <notifications@github.com>
wrote:

> @jonschwenk <https://github.com/jonschwenk> thanks for checking in. The
> review started Jan 13th, and JOSS is requesting reviewers complete their
> reviews within 6 weeks. @eriknes <https://github.com/eriknes> if you
> could complete the review by the middle of next week that would be great.
>
