### 2021-02-27 #2952 @eriknes
I am sorry this took a while. I have installed and tested the software and it works as expected. I have a certain idea about the effort required to make network mapping of river deltas work, so I am actually very impressed!

I have a couple of comments regarding the documentation of the algorithms used and their performance. But it seems like that is not part of the review process? It is the first time I am reviewing for JOSS, and I understand it is a developer oriented journal, but I would have liked to see at least brief descriptions of the algorithms in the software paper, even if references are provided. Skeletonization and pruning methods, as far as I know, usually result in networks with noise.
