### 2021-02-28 #2952 @jonschwenk
Hi @eriknes , glad you were successful in demo-ing RivGraph. It has been challenging and fun to develop, so I'm also glad that you see its value!

As for the algorithm documentation--I believe you are correct in that the purpose of the JOSS paper is not really to explain the algorithm logic, but more to demonstrate their utility. Please note that we do provide full API documentation for (nearly) every function of RG. We've also tried to leave comments throughout to help the curious understand what's going on. For example, take a look at the [prune_delta() source code](https://jonschwenk.github.io/RivGraph/_modules/rivgraph/deltas/delta_utils.html#prune_delta). The steps to prune the vectorized skeleton are laid out fairly explicitly there. If you dig into each of the steps, you'll find similar commenting. We've also tried to name functions as intuitively as possible, often at the expense of brevity. It's not perfect, but we hope it's enough for possible contributors and the curious to unravel the inner workings.
