### 2021-03-01 #2952 @kbarnhart
Sorry for the delay on weighing in.

In a case such as this submission, it would be reasonable to add a bit more to the paper that describe the approach to skeletonization and pruning. Perhaps one more paragraph in "basic functionality" section. That would give the potential user a bit more context about why these methods can be tricky and the approaches that RivGraph takes to address them.

I would also expect that somewhere in the documentation (an overview section, examples, or in the API all seem reasonable places to me) there is sufficient description so that a reader can understand the algorithms conceptually. For example, this part of the delta example: https://jonschwenk.github.io/RivGraph/examples/delta_example/delta_example.html#pruning-the-network

If there are specific algorithmic details that @eriknes thinks should be highlighted in the docs, I'd recommend making an in-repo issue to state what they are and discuss. The balance to keep here is that the algorithm description should serve to help a user understand what the software is doing, and thus use the software correctly.

@jonschwenk Agreed! The comments you've left in the source code are really nice.

Hopefully my comment clarifies these questions, if not, please let me know.
