### 2021-03-03 #2952 @eriknes
Thank you @kbarnhart for weighing in on this

A lot of design choices must have been made to make the output consistent.
Although it may be outside the scope of the paper for this journal to give
the full algorithmic details, I think it would be nice to briefly describe
some of these choices and how well the software can be expected to perform.
For instance, I expect that skeletonization of a river delta network with a
lot of noise along the edges of the channels would get some false positive
branching points / nodes. Maybe the authors could report how well the code
performs compared to a few deltas where the nodes and edges have been
manually delineated, e.g. the ones shown by Tejedor et al.?

Overall a nice piece of work in my opinion

Erik

On Mon, Mar 1, 2021 at 4:43 PM Katy Barnhart <notifications@github.com>
wrote:

> Sorry for the delay on weighing in.
>
> In a case such as this submission, it would be reasonable to add a bit
> more to the paper that describe the approach to skeletonization and
> pruning. Perhaps one more paragraph in "basic functionality" section. That
> would give the potential user a bit more context about why these methods
> can be tricky and the approaches that RivGraph takes to address them.
>
> I would also expect that somewhere in the documentation (an overview
> section, examples, or in the API all seem reasonable places to me) there is
> sufficient description so that a reader can understand the algorithms
> conceptually. For example, this part of the delta example:
> https://jonschwenk.github.io/RivGraph/examples/delta_example/delta_example.html#pruning-the-network
>
> If there are specific algorithmic details that @eriknes
> <https://github.com/eriknes> thinks should be highlighted in the docs,
> I'd recommend making an in-repo issue to state what they are and discuss.
> The balance to keep here is that the algorithm description should serve to
> help a user understand what the software is doing, and thus use the
> software correctly.
>
> @jonschwenk <https://github.com/jonschwenk> Agreed! The comments you've
> left in the source code are really nice.
>
> Hopefully my comment clarifies these questions, if not, please let me know.
>
