### 2021-03-03 #2952 @jonschwenk
We can add some to the text about what the functions are doing. I believe most of this is already in either the examples or the documentation, but I have no problem adding more to the manuscript.

In your example case, @eriknes, the bottom line is that the skeletonization will ~100% respect the mask that you provide. If the mask has many spurious loops (e.g. island "speckling"), those will be retained as RivGraph retains (almost) all connectivity of the mask. If it's just noise on the edges of a channel, these will get pruned if they're not part of a loop. It is up to the user to fill in spurious islands in their mask before applying RivGraph functions. Will clarify this and a few other points in the manuscript and post an update when ready.

Thanks for the feedback!
