### 2021-03-03 9d0cb9d
update paper per JOSS reviews
