### 2021-03-03 7ed051f
pinning gdal to 3.1
Conda GDAL 3.2 does not install correctly; unable to import after installing via environment file. Pinning to 3.1 seems to work fine.
