+++ environment.yml
@@ -5,7 +5,7 @@ channels:
   - defaults
 dependencies:
   - python
-  - gdal
+  - gdal<=3.1
   - numpy
   - geopandas=0.7.0
   - scikit-image
