### 2021-03-03 f5d245c
Update build.yml
CI for more python versions
