### 2021-03-03 72cddfb
Update build.yml
adjust setup-miniconda version
