### 2021-03-03 9aa880c
Update build.yml
back to 3.6
