### 2021-03-03 66f66d8
Update environment.yml
fix python version in environment file
