### 2021-03-04 4104837
Update environment.yml
python 3.8 and gdal 3.1.4
