### 2021-03-04 bbedded
Update build.yml
python 3.8
