### 2021-03-05 #2952 @eriknes
Yes, I think this is probably sufficient as per the JOSS guidelines of
software papers

Thank you

On Thu, Mar 4, 2021 at 4:44 PM Katy Barnhart <notifications@github.com>
wrote:

> @eriknes <https://github.com/eriknes> please let me know if @jonschwenk
> <https://github.com/jonschwenk> 's updates have addressed your concerns.
>
> @jonschwenk <https://github.com/jonschwenk> thanks for making these
> updates.
>
