### 2021-03-05 #2952 @eriknes
Yes, confirmed

On Fri, Mar 5, 2021 at 4:49 PM Katy Barnhart <notifications@github.com>
wrote:

> @leotrs <https://github.com/leotrs> and @eriknes
> <https://github.com/eriknes> based on the comments here, the in-repo
> issues, and the checklists it looks as if your reviews are complete an no
> remaining issues exits. If this is the case and you think this manuscript
> is ready to be accepted, please indicate so here.
>
> @jonschwenk <https://github.com/jonschwenk> once the reviewers have
> indicated the submission is ready to move forward I will complete a couple
> of steps (e.g., copyediting), and then we can move forward with the final
> steps in the JOSS publication process.
>
