### 2021-03-09 e3849c8
Minor copyedits
Minor copyedits associated with the JOSS publication. 

openjournals/joss-reviews/issues/2952
