### 2021-03-09 #2952 @kbarnhart
@jonschwenk I've proposed some very minor copyedits at jonschwenk/RivGraph/pull/48

Please consider, and if you don't disagree, merge. Once you've done that please double check that the paper compiles and complete the following next steps:
