### 2021-03-09 #2952 @jonschwenk
@kbarnhart Checklist complete.

- [x]  Recompiled paper
- [x]  New tagged release has been created: v0.4
- [x]  v0.4 has been stored at Zenodo.
- [x] Metadata matches title and authors, includes ORCIDs.
- [x] Zenodo DOI: 10.5281/zenodo.4591559
