### 2021-03-09 #2952 @kbarnhart
@jonschwenk I've now recommended that is submission be accepted and published. One of the JOSS EICs will handle the final steps @openjournals/joss-eics. Congratulations to you and @elbeejay on a valuable contribution to research software.

Many thanks to reviewers @leotrs and @eriknes for contributing thoughtful and constructive reviews.
